Locate test specific helper code with tests
