+++ TestSwiftyDropbox/TestSwiftyDropbox_iOSTests/ObjCFilesRoutesTests.m
@@ -5,6 +5,12 @@
 #import "ObjCFilesRoutesTests.h"
 #import "ObjCTestClasses.h"
 
+#if TARGET_OS_IPHONE
+#import "TestSwiftyDropbox_iOSTests-Swift.h"
+#elif TARGET_OS_MAC
+#import "TestSwiftyDropbox_macOSTests-Swift.h"
+#endif
+
 
 @implementation ObjCFilesRoutesTests {
     NSOperationQueue *_delegateQueue;
